advisor/pyproject.toml
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[project]
name = "advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "pyyaml>=6.0",
    "openai>=2.8.0,<3",
    "pypandoc>=1.0",
    "nh3>=0.2",
    "lxml>=5.0",
    "tinycss2>=1.0",
    "redis[hiredis]>=5.0",
    "lxml>=6.0.2",
    "tinycss2>=1.5.1",
]
